# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for SBOM microservice",
        "title": "MyESI SBOM Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/sbom",
    "paths": {
        "/github": {
            "post": {
                "description": "Fetch a manifest file from GitHub using REST API and generate a SBOM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM"
                ],
                "summary": "Upload a manifest from a GitHub repository to generate SBOM",
                "parameters": [
                    {
                        "description": "GitHub repo and file info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GitHubSBOMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "Get list of SBOMs filtered by project name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM"
                ],
                "summary": "List SBOMs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Name",
                        "name": "project_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/recent": {
            "get": {
                "description": "Get recent SBOM uploads with optional project filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM"
                ],
                "summary": "Recent SBOM uploads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Name",
                        "name": "project_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Upload a manifest file and generate an SBOM",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM"
                ],
                "summary": "Upload a manifest file to generate SBOM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Name",
                        "name": "project_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Manifest file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "Retrieve a single SBOM by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM"
                ],
                "summary": "Get SBOM by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SBOM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.GitHubSBOMRequest": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "file": {
                    "description": "e.g package.json, go.mod, requirements.txt...",
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                }
            }
        }
    }
}